@@ -1,6 +1,6 @@
 # TL;DR: YOU SHOULD DELETE THIS FILE
 #
-# This file iwas generated by Cucumber-Rails and is only here to get you a head start
+# This file was generated by Cucumber-Rails and is only here to get you a head start
 # These step definitions are thin wrappers around the Capybara/Webrat API that lets you
 # visit pages, interact with widgets and make assertions about page content.
 #
@@ -15,7 +15,7 @@
 #
 # * http://benmabey.com/2008/05/19/imperative-vs-declarative-scenarios-in-user-stories.html
 # * http://dannorth.net/2011/01/31/whose-domain-is-it-anyway/
-# * http://elabs.se/blog/15-you-re-cuking-it-wrong 
+# * http://elabs.se/blog/15-you-re-cuking-it-wrong
 #
 
 
